+++ app/controllers/api/v1/api_controller.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Api
+  module V1
+    class ApiController < Api::ApiController
+    end
+  end
+end
