+++ config/routes.rb
@@ -1,5 +1,9 @@
 # frozen_string_literal: true
 
 Rails.application.routes.draw do
-  resources :users
+  namespace :api do
+    namespace :v1 do
+      resources :users, only: %i[create]
+    end
+  end
 end
